+++ test/vhpi/vhpi_test.c
@@ -1,5 +1,5 @@
 //
-//  Copyright (C) 2023  Nick Gasson
+//  Copyright (C) 2023-2024  Nick Gasson
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU General Public License as published by
@@ -45,6 +45,7 @@ static const vhpi_test_t tests[] = {
    { "vhpi14",   vhpi14_startup },
    { "vhpi15",   vhpi15_startup },
    { "issue978", issue978_startup },
+   { "issue1035", issue1035_startup },
    { NULL,       NULL },
 };