+++ test/vhpi/vhpi_test.h
@@ -48,5 +48,6 @@ void vhpi15_startup(void);
 void issue744_startup(void);
 void issue762_startup(void);
 void issue978_startup(void);
+void issue1035_startup(void);
 
 #endif  // _VHPI_TEST_H